% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rmd_file.R
\name{generate_reports}
\alias{generate_reports}
\title{Build Timor reports}
\usage{
generate_reports(...)
}
\arguments{
\item{...}{List of any additional arguments.}
}
\value{
Documentation on Timor in 3 file formats, as well as the individual
plots.
}
\description{
The function generate Timor reports in 3 formats:
- pdf
- ebook
- gitbook
}
