% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download and generate data}
\usage{
get_data(...)
}
\arguments{
\item{...}{Any additional arguments}
}
\value{
Nothing. The function run the script to upload and generate data
}
\description{
The function run the scripts in `data-raw` folder to download,
process and store the data to be analysed.
}
