% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read configuration file}
\usage{
read_config(conf = "default")
}
\arguments{
\item{conf}{The working environment, choices are `default` or `develop`.}
}
\value{
a list of package configuration attributes
}
\description{
Read configuration file
}
