% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report_data}
\alias{report_data}
\title{Catches and trips from Timor Peskas monitoring system}
\format{
A data frame with 86,128 rows and 16 variables:
}
\usage{
report_data
}
\description{
A dataset containing the catches and trips info of more than 60,000 fishing
trips along Timor coasts.
}
\keyword{datasets}
