% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips_week.R
\docType{data}
\name{trips_per_week}
\alias{trips_per_week}
\title{Estimated number of trips by week}
\format{
A dataframe
}
\usage{
trips_per_week
}
\description{
A dataset containing the estimated number of trips according to the formula
used in `Number of trips per week` section
}
\details{
The column sum indicates the estimated sum of trips per month, while the
column mean indicates the estimated average of trips per month (on a weekly scale).
}
\keyword{datasets}
